% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_GCFAMS.R
\name{predict_GCFAMS}
\alias{predict_GCFAMS}
\title{Calculate Gastric Cancer Fatty Acid Metabolism Score (GCFAMS)}
\usage{
predict_GCFAMS(my_expr)
}
\arguments{
\item{my_expr}{A numeric matrix with rows as genes and columns as samples.}
}
\value{
A data frame with sample names and their Fatty Acid Metabolism Score.
}
\description{
This function calculates the GCFAMS based on a given gene expression matrix.
The expression matrix should have rows as genes and columns as samples.
If the matrix is missing required genes, an error will be thrown.
}
